Causes the event loop to exit on Posix when being interrupted by a signal.
Previously, the loop logic would simply retry the wait until the given timeout was actually reached (or an event got received), which could be forever. This change makes sure that after an interrupt the higher level code will get a chance to run to possibly handle the effects of the signal.

This is a fix for vibe-d/vibe-core#205.